Use smaller integer type in L_targets array
Change integer type to uint16_t, since its range is sufficient for all
possible values (up to kB * kC = 119 * 127 = 15113). This reduces memory
usage and increases performance of finding matches.

+++ src/calculate_bucket.hpp
@@ -68,7 +68,7 @@ std::map<uint8_t, uint8_t> kVectorLens = {
     {8, 0}
 };
 
-uint32_t L_targets[2][kBC][kExtraBitsPow];
+uint16_t L_targets[2][kBC][kExtraBitsPow];
 bool initialized = false;
 void load_tables()
 {
